FM-16: Make storage tests reusable outside the crate
