+++ fendermint/storage/Cargo.toml
@@ -10,13 +10,14 @@ license.workspace = true
 
 [dependencies]
 im = { version = "15.1.0", optional = true }
+quickcheck = { workspace = true, optional = true }
 
 [dev-dependencies]
-quickcheck = { workspace = true }
 quickcheck_macros = { workspace = true }
 serde = { workspace = true }
 fvm_ipld_encoding = { workspace = true }
 
 [features]
-default = ["inmem"]
+default = ["inmem", "testing"]
 inmem = ["im"]
+testing = ["quickcheck"]
